fix: Remove unused files and refactor project structure
Deleted obsolete components, SVGs, and styles to declutter the project. Added new dependencies and updated package files to support API integration and improved modularity.
