+++ api/axios/api-axios.ts
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export const api = axios.create({
+    baseURL: process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3000'
+})
