+++ api/postgres/db.ts
@@ -1,0 +1,11 @@
+import { Pool } from 'pg';
+
+const pool = new Pool({
+    user: '',
+    host: '',
+    database: '',
+    password: '',
+    port: 1234,
+});
+
+export default pool;
