+++ app/page.tsx
@@ -1,0 +1,10 @@
+import {Fragment} from "react";
+import Home from "@/frontend/componentes/template/home";
+
+export default function HomePage() {
+  return (
+      <Fragment>
+          <Home/>
+      </Fragment>
+  );
+}
