+++ package.json
@@ -10,7 +10,10 @@
   },
   "dependencies": {
     "@react-icons/all-files": "^4.1.0",
+    "@types/pg": "^8.11.13",
+    "axios": "^1.8.4",
     "next": "15.3.0",
+    "pg": "^8.14.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0"
