+++ src/frontend/componentes/template/home/home.module.css
@@ -1,0 +1,116 @@
+.homeSection {
+    text-align: center;
+    background: linear-gradient(180deg, #FFF8F0 0%, #FFFDF5 100%);
+    position: relative;
+    overflow: hidden;
+}
+
+/* Bolhas decorativas */
+.homeSection::before {
+    content: '';
+    position: absolute;
+    top: 20%;
+    left: 10%;
+    width: 150px;
+    height: 150px;
+    background-color: rgba(255, 154, 139, 0.08);
+    border-radius: 50%;
+    z-index: 0;
+}
+
+.homeSection::after {
+    content: '';
+    position: absolute;
+    bottom: 10%;
+    right: 5%;
+    width: 200px;
+    height: 200px;
+    background-color: rgba(255, 184, 140, 0.05);
+    border-radius: 50%;
+    z-index: 0;
+}
+
+.homeTitle {
+    font-size: 3rem;
+    font-weight: 700;
+    color: #FF9A8B; /* Coral pastel */
+    margin-bottom: 1.5rem;
+    position: relative;
+    z-index: 1;
+    font-family: 'Pacifico', cursive;
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.05);
+    animation: fadeInDown 1s ease;
+}
+
+.homeSubtitle {
+    font-size: 1.25rem;
+    color: #8B7D6B; /* Marrom pastel */
+    margin-bottom: 2rem;
+    max-width: 600px;
+    margin-left: auto;
+    margin-right: auto;
+    position: relative;
+    z-index: 1;
+    line-height: 1.6;
+}
+
+.homeBanner {
+    border-radius: 16px;
+    box-shadow: 0 8px 25px rgba(255, 154, 139, 0.15);
+    margin: 2rem auto;
+    max-width: 800px;
+    width: 90%;
+    height: auto;
+    display: block;
+    position: relative;
+    z-index: 1;
+    border: 8px solid white;
+    transform: perspective(1000px) rotateX(5deg);
+    transition: all 0.5s cubic-bezier(0.25, 0.8, 0.25, 1);
+}
+
+.homeBanner:hover {
+    transform: perspective(1000px) rotateX(0deg) scale(1.02);
+    box-shadow: 0 12px 30px rgba(255, 154, 139, 0.25);
+}
+
+/* Botão CTA */
+.homeCta {
+    display: inline-block;
+    background-color: #FF9A8B;
+    color: white;
+    padding: 0.8rem 2rem;
+    border-radius: 50px;
+    font-weight: 600;
+    margin-top: 2rem;
+    text-decoration: none;
+    transition: all 0.3s ease;
+    position: relative;
+    z-index: 1;
+    box-shadow: 0 4px 12px rgba(255, 154, 139, 0.2);
+}
+
+.homeCta:hover {
+    background-color: #FFB88C;
+    transform: translateY(-3px);
+    box-shadow: 0 6px 16px rgba(255, 154, 139, 0.3);
+}
+
+@media (max-width: 768px) {
+    .homeSection {
+        padding: 3rem 1rem;
+    }
+
+    .homeTitle {
+        font-size: 2.2rem;
+    }
+
+    .homeSubtitle {
+        font-size: 1.1rem;
+    }
+
+    .homeBanner {
+        width: 95%;
+        border-width: 4px;
+    }
+}
